feat: move response status handling utilily
