+++ planning/mission_planner/package.xml
@@ -20,6 +20,7 @@
   <depend>autoware_adapi_v1_msgs</depend>
   <depend>autoware_auto_mapping_msgs</depend>
   <depend>autoware_planning_msgs</depend>
+  <depend>component_interface_utils</depend>
   <depend>geometry_msgs</depend>
   <depend>lanelet2_extension</depend>
   <depend>motion_utils</depend>
